+++ package.json
@@ -13,13 +13,15 @@
   },
   "dependencies": {
     "@types/node": "20.4.2",
+    "@types/postcss-url": "^10.0.0",
     "@types/react": "18.2.15",
     "@types/react-dom": "18.2.7",
     "autoprefixer": "10.4.14",
     "eslint": "8.45.0",
     "eslint-config-next": "13.4.10",
     "next": "13.4.10",
-    "postcss": "8.4.26",
+    "postcss": "^8.4.26",
+    "postcss-url": "^10.1.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "sass": "^1.63.6",
