+++ src/app/previews/space-tourism-layout/SpaceTourismLayout.tsx
@@ -1,0 +1,5 @@
+import styles from "./SpaceTourismLayout.module.css"
+
+export function SpaceTourismLayout() {
+  return <div className={styles.wrapper}></div>
+}
