+++ src/utils/assetUtil.ts
@@ -1,4 +1,4 @@
-import { assetPrefix, basePath } from "../../next.config";
+import { assetPrefix, basePath } from "../../next.config.js";
 
 /**server only function */
 const getAssetPath = (path: string): string => {
